Verified on the branch by the reviewing session: strict build 0 errors; Native
AOT publish of Loom.AotProbe with no managed dll beside the binary;
790 passed / 0 / 0.

Behaviour change for external listeners: Loom.Telemetry counters are now
published as double rather than long. Loom's own readers are unaffected.

+++ Loom.Telemetry/LoomMetrics.cs
@@ -62,7 +62,7 @@ public static class LoomMetrics
             null
         );
         GetOrCreateBuffer(name).Write(in record);
-        MetricsBridge.PublishCounter(name, (long)value, tags);
+        MetricsBridge.PublishCounter(name, value, tags);
     }
 
     /// <summary>
